+++ CHANGELOG.md
@@ -9,7 +9,8 @@
 * Adding `Norb/MisplacedBranchingLogic` cop to ensure branching logic is not in the wrong place. ([@tleish][])
 * Adding `Norb/ActiveRecordNamespaced` cop to ensure ActiveRecord models are namespaced. ([@tleish][])
 * Adding `Norb/ActiveRecordThroughBusiness` cop to ensure direct ActiveRecord calls come from business objects. ([@tleish][])
-* Adding `Norb/MisplacedLogic` cop to ensure code logic is not in the wrong place. ([@tleish][])
+* Adding `Norb/StandardRestfulControllerActions` cop to ensure code logic is not in the wrong place. ([@tleish][])
+* Adding `Norb/MisplacedLogic` cop checks that Rails Controllers only use standard RESTful actions. ([@tleish][])
 
 
 [@tleish]: https://github.com/tleish